gen update

+++ pages/api/financements/[id].ts
@@ -1,18 +1,31 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { getSession } from "next-auth/react"
 import { JsonApiErrors } from "next-drupal"
+import { Formidable, PersistentFile } from "formidable"
 
 import { drupal } from "lib/drupal"
 
+
+type FormBodyFields = {
+  title: string
+  body: string
+  image: PersistentFile
+}
+
+export const config = {
+  api: {
+    bodyParser: false,
+  },
+}
 export default async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
   try {
-    // Only accept DELETE requests.
-    if (req.method !== "DELETE") {
-      return res.status(405).end()
-    }
+    // // Only accept DELETE requests.
+    // if (req.method !== "DELETE") {
+    //   return res.status(405).end()
+    // }
 
     // Check if the user is authenticated.
     const session = await getSession({ req })
@@ -23,12 +36,53 @@ export default async function handler(
     const id = req.query.id as string
 
     // Delete the article.
-    const deleted = await drupal.deleteResource("node--financement", id, {
-      withAuth: session.accessToken,
-    })
+    if (req.method == "DELETE") {
+      const deleted = await drupal.deleteResource("node--financement", id, {
+        withAuth: session.accessToken,
+      })
 
-    if (!deleted) {
-      throw new Error()
+      if (!deleted) {
+        throw new Error()
+      }
+    }
+
+    // Edit the article.
+    if (req.method == "PUT") {
+
+      const session = await getSession({ req })
+      if (!session) {
+        return res.status(403).end()
+      }
+      const form = new Formidable({
+        keepExtensions: true,
+      })
+
+
+      const fields = await new Promise<FormBodyFields>((resolve, reject) => {
+        form.parse(req, async (error, fields, files) => {
+          if (error) {
+            reject(error)
+            return
+          }
+
+          resolve({
+            title: fields.title,
+
+          })
+        })
+      })
+      const edit = await drupal.updateResource("node--financement", id, {
+        data: {
+          attributes: {
+            title: fields.title,
+               },
+        },
+      },
+    )
+
+      if (!edit) {
+        throw new Error()
+      }
     }
 
     res.end()
@@ -37,6 +91,6 @@ export default async function handler(
       return res.status(error.statusCode).json(error.errors)
     }
 
-    res.status(500).json({ message: "Something went wrong. Please try again." })
+    res.status(500).json({ message: "Something went wrong. truc Please try again." })
   }
 }
